% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcommonality.R
\name{ci_ggcommonality}
\alias{ci_ggcommonality}
\title{Percentile-based Bootstrap Confidence Intervals for ggcommonality.}
\usage{
ci_ggcommonality(
  data.boot,
  formula,
  data,
  ci_sign = "+",
  ci_lower = 0.025,
  ci_upper = 0.975,
  stack_by = "partition",
  ...
)
}
\arguments{
\item{formula}{Formula for linear regression model}

\item{data}{Data frame matching formula argument}

\item{ci_sign}{If "+", genereates confidence intervals using only positive coefficients
If "-", generates confidence intervals using only negative coefficients.
Otherwise, generates confidence interval using both positive and negative.}

\item{ci_lower}{Numeric. Value for lower bound of confidence interval.}

\item{ci_upper}{Numeric. Value for upper bound of confidence interval}

\item{stack_by}{In progress. Currently allows stacking unique and common effects by partition
if "partition" is the input. If stack_by == "common" it stacks unique vs joint effects.}

\item{...}{Additional parameters passed to ggplot2::geom_errorbar}

\item{sample_column}{Character. Name of column for resampling observations. If blank, simple resampling is performed.}

\item{n_replications}{Numeric. Number of replications to use in bootstrap.}

\item{resample_type}{Method for boostrap resampling. Either "random", "fixed", or "wild".}

\item{wild_type}{If resample_type == "wild", either "Gaussian" to
multiply resampled residuals by random constants from the normal distribution,
or sign to randomly multiply half of the residuals by +1 and half by -1.
This provides a solution to "fixed" in the presence of model heteroscedasticity}
}
\value{
Data frame containing commonality partitions for replications.

ggproto instance
}
\description{
Uses structured bootstrapping to create a 95% confidence interval.
}
\details{
Calls yhat::regr() to generate commonality coefficients for
data resampled with replacement. Then it sums unique and joint effects for each
commonality partition and generates a 95% confidence interval.
By setting ci_sign to "positive" or "negative", you can generate an errorbar
for positive and negative commonalities, respectively.
}
