% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterpie-commonality.R
\name{df_commonality_scatterpie}
\alias{df_commonality_scatterpie}
\title{Create tibble for commonality analysis with scatterpie functionality}
\usage{
df_commonality_scatterpie(formula, data)
}
\arguments{
\item{formula}{Formula for linear model from which commonality
coefficients will be extracted.}

\item{data}{Data.frame object for which analysis will be performed.}
}
\value{
Data.frame with proportions for each unique and common
effect.
}
\description{
Create tibble for commonality analysis with scatterpie functionality
}
\details{
The scatterpie package can be useful for colour-coding
commonality coefficients. We can operationalize commonality
coefficients as a scatter pie charts, where the proportion of the pie
is represented by how many cues are involved in the commonality
coefficient. geom_scatterpie requires one column for each effect,
with values representing proportions. We can define a dataframe
such that unique effects receive proportions of 1 in their respective
columns, whereas second-order commonalities receive 0.5 in their
respective columns, etc.

This column prepares a dataframe to simplify commonality plots
when using geom_scatterpie().
}
\examples{
# example with two predictors
df_hp <- df_commonality_scatterpie(
  hp ~ cyl + disp + wt,
  data = mtcars
)
# example with three predictors
df_mpg <- df_commonality_scatterpie(
  mpg ~ cyl + disp + wt,
  data = mtcars
)
# merge them together
df_joined <- dplyr::left_join(df_hp, df_mpg)
# now make scatterpie plot
ggplot2::ggplot() +
 scatterpie::geom_scatterpie(
  ggplot2::aes(
    x = var_hp,
    y = var_mpg
    ),
    data = df_joined,
    cols = c("cyl", "disp", "wt"))+
 ggplot2::coord_fixed()
}
\author{
Cameron Anderson
}
