% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonality_barplot_df.R
\name{df_ggcommonality}
\alias{df_ggcommonality}
\title{Define XY coordinates for drawing commonality bar plots}
\usage{
df_ggcommonality(yhat_model)
}
\arguments{
\item{yhat_model}{Data.frame output of yhat::regr() function.}
}
\value{
List of lists.
Lists for positive and negative commonalities.
Contained are data.frames for drawing barplot [1] effects and [2] outlines.
}
\description{
Define XY coordinates for drawing commonality bar plots
}
\examples{
data(mtcars)
yhat_model_cars <- yhat::regr(
lm(
  formula = mpg ~ cyl + disp + vs,
  data = mtcars
  )
)
df_ggcommonality(yhat_model_cars)
}
