% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterpie-commonality.R
\name{.df_ci_commonality}
\alias{.df_ci_commonality}
\title{Make scatterpie commonality line range plots}
\usage{
.df_ci_commonality(formula, data, interval = c(0.025, 0.975), ...)
}
\arguments{
\item{formula}{Formula for relevant lm object to be bootstrapped}

\item{data}{Data.frame object corresponding to formula}

\item{interval}{Numeric array of 2 values representing confidence
interval to be computed. 95% CIs calculated by default}

\item{...}{Additional parameters passed to
ggcommonality::run_commonality_bootstrap}
}
\value{
Data frame of bootstrapped commonality analysis with
appropriate piechart proportions to represent commonality coefficients
}
\description{
This prepares a dataframe for use with plot_commonality_linerange
(under construction). See [df_2d_commonality()] for details.
}
\author{
Cameron Anderson
}
