% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured-bootstrap.R
\name{.helper_apply_function_to_random_sample}
\alias{.helper_apply_function_to_random_sample}
\title{Apply User-Specified Test Statistic to Resampled Data}
\usage{
.helper_apply_function_to_random_sample(
  dat,
  sample_column,
  samples = 30,
  fnc,
  replace = TRUE,
  ...
)
}
\arguments{
\item{dat}{Data.frame. Data to resample.}

\item{sample_column}{String. ID column containing values to be sampled with replacement.}

\item{samples}{Numeric. Number of samples in return data.frame.}

\item{fnc}{Function. Function to apply to resampled data.}

\item{replace}{Boolean. Sample with replacement?}

\item{...}{Parameters passed to fnc argument.}
}
\value{
Return from function passed to fnc
}
\description{
Uses ellipsis argument to apply a user-specified function to resampled data.
}
\examples{
.helper_apply_function_to_random_sample(dat = mtcars,
sample_column = 'gear',
samples = 30,
fnc = lm,
replace = TRUE,
formula = mpg ~ cyl)

}
