% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured-bootstrap.R
\name{.helper_apply_test_statistic}
\alias{.helper_apply_test_statistic}
\title{Generic Function for Test Statistic}
\usage{
.helper_apply_test_statistic(fnc, ...)
}
\arguments{
\item{fnc}{name of function to use in bootstrap}

\item{...}{Optional parameters for fnc}
}
\value{
Output of test function. See details for corresponding help page.
}
\description{
Generic Function for Test Statistic
}
\examples{
.helper_apply_test_statistic(fnc = lm,
data = mtcars,
formula = mpg ~ cyl)

}
