% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessers.R
\name{.helper_define_x_coordinates}
\alias{.helper_define_x_coordinates}
\title{Define X and Y coordinates for building commonality barplot}
\usage{
.helper_define_x_coordinates(pivoted_cue_df, unpivoted_cue_df, x_offset = 1.5)
}
\arguments{
\item{pivoted_cue_df}{Data.frame output from .helper_duplicate_inner_values}

\item{unpivoted_cue_df}{Data.frame output of .helper_split_partition_effects}

\item{x_offset}{Numeric. How much to offset adjacent partitions?}
}
\value{
Data.frame object containing x coordinates for commonality.
}
\description{
Separate adjacent commonality partitions.
}
