% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessers.R
\name{.helper_define_y_coordinates}
\alias{.helper_define_y_coordinates}
\title{Define X and Y coordinates for building commonality barplot}
\usage{
.helper_define_y_coordinates(pivoted_cue_df)
}
\arguments{
\item{pivoted_cue_df}{Data.frame output from .helper_define_x_coordinates}
}
\value{
Data.frame containing x a y coordinates for drawing commonalities.
}
\description{
Incrementally add common effects on top of unique effects.
}
