% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessers.R
\name{.helper_draw_barplot_outline}
\alias{.helper_draw_barplot_outline}
\title{Define X and Y coordinates for building commonality barplot}
\usage{
.helper_draw_barplot_outline(pivoted_cue_df_xy, type = "positive")
}
\arguments{
\item{pivoted_cue_df_xy}{Data.frame. Output of.helper_define_y_coordinates}

\item{type}{Positive or negative.
Required for plotting positive and negative barplot effects.}
}
\value{
Data.frame object containing outline for commonality bar plot.
}
\description{
Get coordinates for drawing outline of commonality barplot
}
