% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-level-helpers.R
\name{.helper_duplicate_inner_values}
\alias{.helper_duplicate_inner_values}
\title{Simple Helper Functions for Commonality Barplot}
\usage{
.helper_duplicate_inner_values(nums)
}
\arguments{
\item{nums}{Numeric array.}
}
\value{
Array.Returns values in nums, with inner values duplicated.
}
\description{
Function used to duplicate rows by number of variables.
Used when defining common effect bar width.
}
\details{
Duplicate the inner values of a sequence.
}
