% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessers.R
\name{.helper_format_yhat_commonality}
\alias{.helper_format_yhat_commonality}
\title{Preprocessing Functions for Commonality Barplot}
\usage{
.helper_format_yhat_commonality(fitted_model, type = "positive")
}
\arguments{
\item{fitted_model}{List returned by yhat::regr() function.}

\item{type}{Character. One of "positive" or "negative".
Return positive and negative commonality values from model respectively}
}
\value{
Data.frame object of commonality effects.
Formatted in data.frame compatible with commonalitybarplot functions.
}
\description{
Format yhat::regr() output for plotting purposes.
}
