% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessers.R
\name{.helper_pivot_cues_longer}
\alias{.helper_pivot_cues_longer}
\title{Preprocessing Functions for Commonality Barplot}
\usage{
.helper_pivot_cues_longer(partition_effect_df)
}
\arguments{
\item{partition_effect_df}{Data.frame returned from .helper_split_partition_effects}
}
\value{
Data.frame with pivoted values.
}
\description{
Pivot commonality data.frame longer.
}
