% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured-bootstrap.R
\name{.helper_resample_df}
\alias{.helper_resample_df}
\title{Resample Data with Replacement by ID Column}
\usage{
.helper_resample_df(dat, sample_column, samples = 30, replace = TRUE)
}
\arguments{
\item{dat}{Data.frame object.}

\item{sample_column}{String. ID column containing values to be sampled with replacement.}

\item{samples}{Numeric. Number of samples in return data.frame.}

\item{replace}{Boolean. Sample with replacement?}
}
\value{
Data.frame. Resampled data.
}
\description{
Resample Data with Replacement by ID Column
}
\examples{
.helper_resample_df(mtcars,
sample_column = 'gear',
samples = 30,
replace = TRUE)
}
\author{
Cameron Anderson
}
