% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-level-helpers.R
\name{.helper_seq_from_1_to_2_by}
\alias{.helper_seq_from_1_to_2_by}
\title{Simple Helper Functions for Commonality Barplot}
\usage{
.helper_seq_from_1_to_2_by(x)
}
\arguments{
\item{x}{Numeric. Value or array of values to sequence across.}
}
\value{
List. For each value in x, returns a sequence from 1 to 2 by x.
}
\description{
Function used for defining the width of individual and common effects
}
\details{
Sequence from 1 to 2 by a user-input name.
Function applied across an array of values.
}
