% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.helper_simplify_df}
\alias{.helper_simplify_df}
\title{Drop columns unnecessary to bootstrap commonality analysis}
\usage{
.helper_simplify_df(formula, data, groups)
}
\arguments{
\item{formula}{Formula to be used in helper_commonality_bootstrap}

\item{data}{Data to be used in helper_commonality_bootstrap}

\item{groups}{Groups to be used in helper_commonality_bootstrap}
}
\value{
Data.frame. Drops columns that don't include terms in formula
}
\description{
Gets rid of columns not specified in formula
}
