% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessers.R
\name{.helper_split_partition_effects}
\alias{.helper_split_partition_effects}
\title{Preprocessing Functions for Commonality Barplot}
\usage{
.helper_split_partition_effects(plot_data, splitter = ",", n_pairs)
}
\arguments{
\item{plot_data}{Data.frame with columns "names" and "values".
Names column should include a consistent splitter between variable names.}

\item{splitter}{String pattern used to separate joint effects. E.g., "and"}

\item{n_pairs}{The number of variable splits to look for.}
}
\value{
Data.frame object.
}
\description{
Expand out data frame so common effects are represented in separate rows.
Use regex functions for splitting.
}
