% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{funs}
\alias{funs}
\alias{sum1}
\alias{prod1}
\title{Two functions of sum1 and prod1}
\usage{
sum1(x, y)

prod1(x, y)
}
\arguments{
\item{x}{=X}

\item{y}{=Y}
}
\value{
x*y (prod1) or x+y (sum1).
}
\description{
Two functions of sum1 and prod1
}
\examples{
sum1(3,4)
prod1(3,4)
}
