% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonality_barplot_df.R
\name{get_commonality_barplot_df}
\alias{get_commonality_barplot_df}
\title{Define XY coordinates for drawing commonality bar plots}
\usage{
get_commonality_barplot_df(yhat_model)
}
\arguments{
\item{yhat_model}{Data.frame output of yhat::regr() function.}
}
\value{
List of lists.
Lists for positive and negative commonalities.
Contained are data.frames for drawing barplot [1] effects and [2] outlines.
}
\description{
Define XY coordinates for drawing commonality bar plots
}
