% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{ggcom}
\alias{ggcom}
\title{Object for plotting commonality analyses}
\usage{
ggcom(
  formula,
  data,
  add_ci = TRUE,
  stack = NULL,
  n_replications = 100,
  sample_column = NULL,
  resample_type = "wild",
  wild_type = "gaussian",
  include_total = FALSE,
  ci_bounds = c(0.025, 0.975),
  seed = NULL
)
}
\arguments{
\item{formula}{Formula in form of y ~ x1 + x2}

\item{data}{Data.frame object containing data to be visualized}

\item{add_ci}{Boolean. Add confidence interval generated from bootstrapping?}

\item{stack}{Character specifying how to stack commonality coefficients. Either NULL for no stacking, "common" to stack unique vs. common effects or "partition" to stack by commonality partition.}

\item{n_replications}{Numeric. Number of replications for bootstrap simulation.}

\item{sample_column}{Character. Name of column to perform stratified sampling with, or leave as NULL}

\item{resample_type}{Character. Method for boostrap resampling. Either "random", "fixed", or "wild". See README for details.}

\item{wild_type}{Character. If resample_type == "wild", either "Gaussian" to
multiply resampled residuals by random constants from the normal distribution,
or sign to randomly multiply half of the residuals by +1 and half by -1.
This provides a solution to "fixed" in the presence of model heteroscedasticity. See README for details}

\item{include_total}{Logical. Include bar representing total variance explained across all unique and common effects?}

\item{ci_bounds}{Array. Values representing lower and upper bounds of confidence intervals.}

\item{seed}{Numeric. Number to set R's randomization seed to for reproducibility.}
}
\description{
Function for defining ggcommonality object
}
