% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{ggcom_ci}
\alias{ggcom_ci}
\title{Flexibly plots confidence interval for GGCommonality objects based on percentile-based bootstrapping using ggplot2}
\usage{
ggcom_ci(x, ...)
}
\arguments{
\item{x}{GGCommonality object}

\item{...}{Other ggplot2 parameters}
}
\description{
Flexibly plots confidence interval for GGCommonality objects based on percentile-based bootstrapping using ggplot2
}
