% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{ggcom_yhat}
\alias{ggcom_yhat}
\title{Get results from yhat package's commonality analysis function}
\usage{
ggcom_yhat(x)
}
\arguments{
\item{x}{GGCommonality object}
}
\value{
ggproto object
}
\description{
Results from yhat for commonality analysis.
}
