% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{class}
\name{GGCommonality-class}
\alias{GGCommonality-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame.}

\item{\code{data.boot}}{matrix.}

\item{\code{formula}}{formula.}

\item{\code{stack}}{ANY.}

\item{\code{stack_by}}{character.}

\item{\code{n_replications}}{numeric.}

\item{\code{sample_column}}{ANY.}

\item{\code{resample_type}}{character.}

\item{\code{wild_type}}{character.}

\item{\code{include_total}}{ANY.}

\item{\code{seed}}{ANY.}

\item{\code{...}}{ANY.}
}}

