% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcommonality.R
\name{ggcommonality}
\alias{ggcommonality}
\title{Plot Commonality Partitions as a Bar Plot in ggplot}
\usage{
ggcommonality(yhat_model)
}
\arguments{
\item{yhat_model}{List. Model output of yhat::reg() applied to lm() object.}
}
\value{
ggplot object. Unique and common effects presented as a bar plot.
Variance attributable to two variables appears in partition for both.
}
\description{
This function takes a commonality model output from yhat and generates
a bar plot based on the unique and common variance shared between variables.
}
\examples{
data(mtcars)
yhat_model <- yhat::regr(
lm(
  formula = mpg ~ cyl + disp + vs + drat + hp + wt + am + gear +carb,
  data = mtcars
  )
)
ggcommonality(yhat_model)

}
