% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcommonality.R
\name{ggcommonality_ci}
\alias{ggcommonality_ci}
\title{Generate percentile-based bootstrap 95% confidence intervals for ggcommonality}
\usage{
ggcommonality_ci(formula, data, ...)
}
\arguments{
\item{formula}{Formula for linear regression model}

\item{data}{Data frame matching formula argument}

\item{...}{Additional parameters passed to [helper_calc_ci()]}
}
\value{
ggproto instance
}
\description{
Generate percentile-based bootstrap 95% confidence intervals for ggcommonality
}
\examples{
ggcommonality(formula = mpg ~ cyl + disp + vs,
data = mtcars) +
ggcommonality_ci(formula = mpg ~ cyl + disp + vs,
data = mtcars,
sample_column = gear)
}
