% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{plot,GGCommonality-method}
\alias{plot,GGCommonality-method}
\alias{plot,ggcommonality}
\alias{ggcom_ci,GGCommonality-method}
\alias{plot,ggcommonality,missing-method}
\title{Plot commonality analyses}
\usage{
\S4method{plot}{GGCommonality}(x)

\S4method{ggcom_ci}{GGCommonality}(x, width = 0.3, ...)
}
\arguments{
\item{x}{A GGCommonality class object}

\item{width}{Width argument passed to ggplot2 to define confidence interval appearance}

\item{...}{Other arguments passed to ggprotos from ggplot2}
}
\description{
Plots commonality analysis confidence intervals from bootstrap simulations

This method plots a ggcommonality object
}
