% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{plot_commonality}
\alias{plot_commonality}
\title{Object for plotting commonality analyses}
\usage{
plot_commonality(
  formula,
  data,
  add_ci = TRUE,
  stack = FALSE,
  stack_by = "common",
  n_replications = 100,
  sample_column = NULL,
  resample_type = "wild",
  wild_type = "gaussian",
  include_total = FALSE,
  seed = NULL
)
}
\arguments{
\item{formula}{Formula in form of y ~ x1 + x2}

\item{data}{Data.frame object containing data to be visualized}

\item{add_ci}{Logical. Add bootstrap-estimated confidence interval?}

\item{stack}{Logical. Stack commonality effects?}

\item{stack_by}{Character. Either "common" to stack unique vs. common effects or "partition" to stack unique and common effects for each IV.}

\item{n_replications}{Numeric. Number of replications for bootstrap simulation.}

\item{sample_column}{Character. Name of column to perform stratified sampling with, or leave as NULL}

\item{resample_type}{Character. Method for boostrap resampling. Either "random", "fixed", or "wild". See README for details.}

\item{wild_type}{Character. If resample_type == "wild", either "Gaussian" to
multiply resampled residuals by random constants from the normal distribution,
or sign to randomly multiply half of the residuals by +1 and half by -1.
This provides a solution to "fixed" in the presence of model heteroscedasticity. See README for details}

\item{include_total}{Logical. Include bar representing total variance explained across all unique and common effects?}

\item{seed}{Numeric. Number to set R's randomization seed to (for reproducibility).}
}
\description{
Object for plotting commonality analyses
}
