% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcommonality.R
\name{plot_ggcommonality}
\alias{plot_ggcommonality}
\title{Plot Commonality Partitions as a Bar Plot in ggplot}
\usage{
plot_ggcommonality(formula, data, stack = "partition")
}
\arguments{
\item{formula}{Formula passed to regression model}

\item{data}{data argument matching formula}

\item{stack}{Character specifying how to stack commonality coefficients. Either NULL for no stacking, "common" to stack unique vs. common effects or "partition" to stack by commonality partition.}
}
\value{
ggplot object. Unique and common effects presented as a bar plot.
Variance attributable to two variables appears in partition for both.
}
\description{
This function takes a commonality model output from yhat and generates
a bar plot based on the unique and common variance shared between variables.
}
\examples{
data(mtcars)
plot_ggcommonality(formula = mpg ~ hp + wt, data = mtcars) |>
  suppressWarnings()
}
