% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{run_commonality_bootstrap}
\alias{run_commonality_bootstrap}
\title{Perform bootstrapping on commonality coefficients}
\usage{
run_commonality_bootstrap(
  formula,
  data,
  groups,
  resample_type = "random",
  n_replications = 10000
)
}
\arguments{
\item{formula}{Formula to be used in helper_commonality_bootstrap}

\item{data}{Data to be used in helper_commonality_bootstrap}

\item{groups}{Groups to be used in helper_commonality_bootstrap}

\item{resample_type}{Method for boostrap resampling. Either "random" or "fixed"}

\item{n_replications}{Number of replications in bootstrap}
}
\value{
Data frame containing commonality partitions for replications.
}
\description{
Get commonality coefficients for n replications
}
\examples{
run_commonality_bootstrap(formula = cyl ~ mpg + hp,
data = mtcars,
groups = "gear",
n_replications = 100) |> suppressWarnings()
}
