% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured-bootstrap.R
\name{run_structured_bootstrap}
\alias{run_structured_bootstrap}
\title{Bootstrap Data Using User-Specified Function}
\usage{
run_structured_bootstrap(
  dat,
  sample_column,
  samples = 30,
  replications = 100,
  fnc,
  replace = TRUE,
  ...
)
}
\arguments{
\item{dat}{Data.frame. Data to resample.}

\item{sample_column}{String. ID column containing values to be sampled with replacement.}

\item{samples}{Numeric. Number of samples in return data.frame.}

\item{replications}{Numeric. Number of times to replicate analysis.}

\item{fnc}{Function. Function to apply to resampled data.}

\item{replace}{Boolean. Sample with replacement?}

\item{...}{Parameters passed to fnc argument.}
}
\value{
Data.frame object containing output of fnc
}
\description{
Perform a structured bootstrap with a user-specified function
}
\examples{
# Example 1: Bootstrapping linear regression
boot_lm <- run_structured_bootstrap(dat = mtcars,
sample_column = 'gear',
samples = 30,
replications = 100,
fnc = lm,
replace = TRUE,
formula = mpg ~ cyl)
print(boot_lm[,1:2])

# Example 2: User-created function
# load in external library (for example)
library(yhat)
# create a user-specified function to apply in bootstrap
commonality_analysis <- function(...) {
yhat::regr(lm(...))
}
# perform bootstrap
boot_com <- run_structured_bootstrap(dat = mtcars,
sample_column = 'gear',
samples = 30,
replications = 100,
fnc = commonality_analysis,
replace = TRUE,
formula = mpg ~ cyl + am)
print(boot_com[,1:2])
}
\author{
Cameron Anderson
}
