% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcommonality.R
\name{ci_ggcommonality}
\alias{ci_ggcommonality}
\title{Percentile-based Bootstrap Confidence Intervals for ggcommonality.}
\usage{
ci_ggcommonality(formula, data, ...)
}
\arguments{
\item{formula}{Formula for linear regression model}

\item{data}{Data frame matching formula argument}

\item{...}{Additional parameters passed to helper_calc_ci()}
}
\value{
ggproto instance
}
\description{
Uses structured bootstrapping to create a 95% confidence interval.
}
\details{
Calls yhat::regr() to generate commonality coefficients for
data resampled with replacement. Then it sums unique and joint effects for each
commonality partition and generates a 95% confidence interval.
By setting ci_sign to "positive" or "negative", you can generate an errorbar
for positive and negative commonalities, respectively.
}
\examples{
 ggcommonality(formula = mpg ~ cyl + disp + vs,
               data = mtcars) +
   ci_ggcommonality(formula = mpg ~ cyl + disp + vs,
                    data = mtcars,
                    sample_column = "gear")
}
