% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonality_barplot_df.R
\name{df_ggcommonality}
\alias{df_ggcommonality}
\title{Define XY coordinates for drawing commonality bar plots}
\usage{
df_ggcommonality(formula, data, stack = "partition")
}
\arguments{
\item{formula}{Formula passed to regression model}

\item{data}{data argument matching formula}

\item{stack}{Character specifying how to stack commonality coefficients. Either NULL for no stacking, "common" to stack unique vs. common effects or "partition" to stack by commonality partition.
Lists for positive and negative commonalities.
Contained are data.frames for drawing barplot [1] effects and [2] outlines.}
}
\description{
Define XY coordinates for drawing commonality bar plots
}
\examples{
data(mtcars)
df_ggcommonality(formula = mpg ~ cyl + disp + vs, data = mtcars) |>
  suppressWarnings()
}
