% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonality_barplot_df.R
\name{df_ggcommonality}
\alias{df_ggcommonality}
\title{Define XY coordinates for drawing commonality bar plots}
\usage{
df_ggcommonality(formula, data, by = "partition")
}
\arguments{
\item{formula}{Formula passed to regression model}

\item{data}{data argument matching formula}

\item{by}{In progress. Currently allows stacking unique and common effects by partition
if "partition" is the input. Otherwise it stacks unique vs joint effects.}
}
\value{
List of lists.
Lists for positive and negative commonalities.
Contained are data.frames for drawing barplot [1] effects and [2] outlines.
}
\description{
Define XY coordinates for drawing commonality bar plots
}
\examples{
data(mtcars)
df_ggcommonality(formula = mpg ~ cyl + disp + vs, data = mtcars) |>
  suppressWarnings()
}
