% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-ci.R
\name{.helper_make_ci}
\alias{.helper_make_ci}
\title{Generate percentile-based confidence intervals for commonality coefficients}
\usage{
.helper_make_ci(
  data,
  formula,
  ci_sign = "+",
  ci_lower = 0.025,
  ci_upper = 0.975,
  stack_by = "partition"
)
}
\arguments{
\item{data}{Data to sample observations from}

\item{formula}{Formula corresponding to linear regression model}

\item{ci_sign}{Character. Sign corresponding to which coefficients should be used for generating error bar for confidence interval. If sign = "+", samples only positive coefficients; if "-", only negative coefficients.}

\item{ci_lower}{Lower bound of confidence interval.}

\item{ci_upper}{Upper bound of confidence interval.}

\item{stack_by}{If "partition", samples from unique and joint effects for commonality partition. If "common", creates confidence interval based on unique vs. common effects.}

\item{sample_column}{Optional column to sample from}

\item{resample_type}{Method for boostrap resampling. Either "random", "fixed", or "wild".}

\item{wild_type}{If resample_type == "wild", either "Gaussian" to
multiply resampled residuals by random constants from the normal distribution,
or sign to randomly multiply half of the residuals by +1 and half by -1.
This provides a solution to "fixed" in the presence of model heteroscedasticity}

\item{n_replications}{The number of replications to perform in bootstrap simulation.}
}
\value{
Data frame containing commonality partitions for replications.

Data.frame object containing confidence intervals for each variable.
}
\description{
Create random- or fixed-effect percentile-based bootstrap intervals.
}
\author{
Cameron Anderson, Julianne Heitelman
}
