% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{.helper_resample_commonality}
\alias{.helper_resample_commonality}
\title{Commonality Coefficients From Resampled Data}
\usage{
.helper_resample_commonality(
  formula,
  data,
  groups,
  resample_type,
  wild_type = "gaussian"
)
}
\arguments{
\item{formula}{Formula to be used in helper_commonality_bootstrap}

\item{data}{Data to be used in helper_commonality_bootstrap}

\item{groups}{Groups to be used in helper_commonality_bootstrap}

\item{resample_type}{Character vector specifying whether resampling should be fixed or random. See details.}
}
\value{
Vector of commonality coefficients on resampled data.
}
\description{
Calculate commonality coefficients in bootstrap procedure
}
