% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{class}
\name{GGCommonality-class}
\alias{GGCommonality-class}
\title{Visualizing commonality analyses}
\value{
Data frame containing commonality partitions for replications.

List with results from yhat::regr commonality analysis summary
}
\description{
S4 class for defining ggcommonality analysis objects.
}
\section{Slots}{

\describe{
\item{\code{data}}{Data.frame object containing data to be visualized.}

\item{\code{data.boot}}{Matrix of bootstrapped data used to generate confidence interval.}

\item{\code{formula}}{formula. Formula representing equation for linear regression model.}

\item{\code{stack}}{Character specifying how to stack commonality coefficients. Either NULL for no stacking, "common" to stack unique vs. common effects or "partition" to stack by commonality partition.}

\item{\code{n_replications}}{numeric. Number of bootstrap replications.}

\item{\code{sample_column}}{ANY. Column to resample within in bootstrap (see mosaic::resample()).}

\item{\code{resample_type}}{character. Character vector specifying whether resampling should be fixed or random. See details.}

\item{\code{wild_type}}{character.If resample_type == "wild", either "Gaussian" to
multiply resampled residuals by random constants from the normal distribution,
or sign to randomly multiply half of the residuals by +1 and half by -1.
This provides a solution to "fixed" in the presence of model heteroscedasticity}

\item{\code{include_total}}{ANY. TRUE or FALSE, specifying whether to include bar representing total explained variance.}

\item{\code{seed}}{ANY. Number specifying which seed to set R's random number generator to.}

\item{\code{ci_bounds}}{ANY.}

\item{\code{...}}{ANY.}
}}

