% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcommonality.R
\name{ggcommonality}
\alias{ggcommonality}
\title{Plot Commonality Partitions as a Bar Plot in ggplot}
\usage{
ggcommonality(formula, data, by = "partition")
}
\arguments{
\item{formula}{Formula passed to regression model}

\item{data}{data argument matching formula}

\item{by}{In progress. Currently allows stacking unique and common effects by partition
if "partition" is the input. Otherwise it stacks unique vs joint effects.}
}
\value{
ggplot object. Unique and common effects presented as a bar plot.
Variance attributable to two variables appears in partition for both.
}
\description{
This function takes a commonality model output from yhat and generates
a bar plot based on the unique and common variance shared between variables.
}
\examples{
data(mtcars)
ggcommonality(formula = mpg ~ cyl + disp + vs, data = mtcars) |>
  suppressWarnings()

data(trees)
ggcommonality(formula = Height ~ Girth + Volume + Girth * Volume,
data = trees)
}
