% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{run_commonality_bootstrap}
\alias{run_commonality_bootstrap}
\title{Perform bootstrapping on commonality coefficients}
\usage{
run_commonality_bootstrap(
  formula,
  data,
  groups = NULL,
  resample_type = "random",
  wild_type = "gaussian",
  n_replications = 100,
  seed = NULL
)
}
\arguments{
\item{formula}{Formula to be used in helper_commonality_bootstrap}

\item{data}{Data to be used in helper_commonality_bootstrap}

\item{groups}{Groups to be used in helper_commonality_bootstrap}

\item{resample_type}{Method for boostrap resampling. Either "random", "fixed", or "wild".}

\item{wild_type}{If resample_type == "wild", either "Gaussian" to
multiply resampled residuals by random constants from the normal distribution,
or sign to randomly multiply half of the residuals by +1 and half by -1.}

\item{n_replications}{Number of replications in bootstrap}

\item{seed}{Numeric. Value to set R's random number generator to for reproducibility
This provides a solution to "fixed" in the presence of model heteroscedasticity}
}
\value{
Data frame containing commonality partitions for replications.
}
\description{
Get commonality coefficients for n replications
}
\examples{
run_commonality_bootstrap(formula = cyl ~ mpg + hp,
data = mtcars,
groups = "gear",
n_replications = 100) |> suppressWarnings()
}
